refactor: MainPageLayout 변경 - MainLayoutComponents 폴더를 만들고 그 안에 사이드바와 상단바를 컴포넌트화 시켜서 분리
